+++ zarrs_plugin/CHANGELOG.md
@@ -7,6 +7,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.3.1] - 2026-01-09
+
 ### Fixed
 - Use `paste` re-export in the `impl_extension_aliases` macro
 
@@ -73,7 +75,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
  - Initial release
  - Split from the `plugin` module of `zarrs` 0.20.0-dev
 
-[unreleased]: https://github.com/zarrs/zarrs/compare/zarrs_plugin-v0.3.0...HEAD
+[unreleased]: https://github.com/zarrs/zarrs/compare/zarrs_plugin-v0.3.1...HEAD
+[0.3.1]: https://github.com/zarrs/zarrs/releases/tag/zarrs_plugin-v0.3.1
 [0.3.0]: https://github.com/zarrs/zarrs/releases/tag/zarrs_plugin-v0.3.0
 [0.2.3]: https://github.com/zarrs/zarrs/releases/tag/zarrs_plugin-v0.2.3
 [0.2.2]: https://github.com/zarrs/zarrs/releases/tag/zarrs_plugin-v0.2.2